+++ src/array.mjs
@@ -67,3 +67,4 @@ export { default as uksort } from './array/uksort.mjs';
 export { default as uasort } from './array/uasort.mjs';
 export { default as natsort } from './array/natsort.mjs';
 export { default as natcasesort } from './array/natcasesort.mjs';
+export { default as ksort } from './array/ksort.mjs';