+++ src/array/ksort.mjs
@@ -1,0 +1,32 @@
+import { SORT_USE_KEY } from './constants.mjs';
+import multisort from './multisort.mjs';
+
+/**
+ * Сортирует массив по ключу в порядке возрастания
+ *
+ * Функция сортирует массив `array` по ключу в порядке возрастания.
+ *
+ * > **Замечание:**
+ * >
+ * > Функция сохраняет первоначальный порядок элементов, если при сравнении значения двух элементов равны.
+ *
+ * @param {Array|Object} array Входной массив.
+ * @param {Number} flags Необязательный второй параметр `flags` изменяет поведение сортировки
+ * и может принимать следующие значения:
+ *
+ * Флаги типов сортировки:
+ *
+ * - `SORT_REGULAR` — обычное сравнение элементов
+ * - `SORT_NUMERIC` — числовое сравнение элементов
+ * - `SORT_STRING` — строковое сравнение элементов
+ * - `SORT_NATURAL` — сравнение элементов как строки,
+ * используя "естественный порядок", например `natsort()`
+ * - `SORT_FLAG_CASE` — можно объединять (побитовое ИЛИ) с `SORT_STRING`
+ * или `SORT_NATURAL` для сортировки строк без учёта регистра
+ *
+ * @returns {true} Функция всегда возвращает `true`.
+ * @throws {TypePHPJSError|ValuePHPJSError}
+ */
+export default function ksort(array, flags) {
+    return multisort(array, flags, SORT_USE_KEY);
+}